feat(themes): add omarchy themes

+++ bin/omakupt-sub/theme.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
-THEME_NAMES=("Tokyo Night" "Catppuccin" "Nord" "Everforest" "Gruvbox" "Kanagawa" "Ristretto" "Rose Pine" "Matte Black" "Osaka Jade")
-THEME=$(gum choose "${THEME_NAMES[@]}" "<< Back" --header "Choose your theme" --height 12 | tr '[:upper:]' '[:lower:]' | sed 's/ /-/g')
+THEME_NAMES=("Tokyo Night" "Catppuccin" "Catppuccin Latte" "Nord" "Everforest" "Gruvbox" "Kanagawa" "Ristretto" "Rose Pine" "Matte Black" "Osaka Jade" "Ethereal" "Flexoki Light" "Hackerman" "Last Horizon" "Lumon" "Lupine" "Miasma" "Retro 82" "Solitude" "Vantablack" "White")
+THEME=$(gum choose "${THEME_NAMES[@]}" "<< Back" --header "Choose your theme" --height 24 | tr '[:upper:]' '[:lower:]' | sed 's/ /-/g')
 
 if [ -n "$THEME" ] && [ "$THEME" != "<<-back" ]; then
   cp $OMAKUPT_PATH/themes/$THEME/alacritty.toml ~/.config/alacritty/theme.toml
   cp $OMAKUPT_PATH/themes/$THEME/zellij.kdl ~/.config/zellij/themes/$THEME.kdl
   sed -i "s/theme \".*\"/theme \"$THEME\"/g" ~/.config/zellij/config.kdl
-  if [ -d "$HOME/.config/nvim" ]; then
+  if [ -d "$HOME/.config/nvim" ] && [ -f "$OMAKUPT_PATH/themes/$THEME/neovim.lua" ]; then
     cp $OMAKUPT_PATH/themes/$THEME/neovim.lua ~/.config/nvim/lua/plugins/theme.lua
   fi
 
@@ -20,7 +20,7 @@ if [ -n "$THEME" ] && [ "$THEME" != "<<-back" ]; then
 
   source $OMAKUPT_PATH/themes/$THEME/gnome.sh
   source $OMAKUPT_PATH/themes/$THEME/tophat.sh
-  source $OMAKUPT_PATH/themes/$THEME/vscode.sh
+  [ -f "$OMAKUPT_PATH/themes/$THEME/vscode.sh" ] && source $OMAKUPT_PATH/themes/$THEME/vscode.sh
 
   # Forgo setting the Chrome theme until we might find a less disruptive way of doing it.
   # Having to quit Chrome, and all Chrome-based apps, is too much of an inposition.
